refactor(management): centralize request mapping
